refactor(js): defer standalone text entry
