+++ packages/js/build.config.ts
@@ -4,7 +4,7 @@ export default defineBuildConfig({
   entries: [
     {
       type: 'bundle',
-      input: ['./src/index.ts', './src/negotiate.ts', './src/plugins.ts', './src/preset/minimal.ts', './src/splitter.ts', './src/parse.ts', './src/text.ts', './src/cli.ts', './src/llms-txt.ts'],
+      input: ['./src/index.ts', './src/negotiate.ts', './src/plugins.ts', './src/preset/minimal.ts', './src/splitter.ts', './src/parse.ts', './src/cli.ts', './src/llms-txt.ts'],
     },
   ],
 })
