+++ packages/js/src/markdown-processor.ts
@@ -1,4 +1,4 @@
-import type { ParseState } from './parse-core'
+import type { ParseState } from './parse'
 import type { ElementNode, EngineOptions, HandlerContext, NodeEvent, PluginContext, TagHandler, TextNode, TransformPlugin } from './types'
 import {
   DEFAULT_BLOCK_SPACING,
@@ -26,9 +26,8 @@ import {
   TAG_TH,
   TEXT_NODE,
 } from './const'
-import { finalizeParse, parseHtmlStream } from './parse-core'
+import { finalizeParse, parseHtmlStream } from './parse'
 import { processPluginsForEvent } from './plugin-processor'
-import { tagHandlers } from './tags'
 
 export interface MarkdownState {
   /** Configuration options for conversion */
@@ -625,7 +624,6 @@ export function createMarkdownProcessor(options: EngineOptions = {}, resolvedPlu
       depth: 0,
       resolvedPlugins,
       tagOverrideHandlers,
-      tagHandlers,
       plainText: state.plainText,
     }
