--- packages/js/src/tag-metadata.ts
@@ -1,201 +1,0 @@
-import type { TagHandler } from './types'
-import {
-  NO_SPACING,
-  TABLE_ROW_SPACING,
-  TAG_A,
-  TAG_ABBR,
-  TAG_AREA,
-  TAG_B,
-  TAG_BASE,
-  TAG_BDO,
-  TAG_BR,
-  TAG_BUTTON,
-  TAG_CITE,
-  TAG_CODE,
-  TAG_COL,
-  TAG_DATALIST,
-  TAG_DEL,
-  TAG_DFN,
-  TAG_EM,
-  TAG_EMBED,
-  TAG_FIGCAPTION,
-  TAG_HEAD,
-  TAG_I,
-  TAG_IFRAME,
-  TAG_IMG,
-  TAG_INPUT,
-  TAG_INS,
-  TAG_KBD,
-  TAG_KEYGEN,
-  TAG_LABEL,
-  TAG_LINK,
-  TAG_MARK,
-  TAG_META,
-  TAG_NOFRAMES,
-  TAG_NOSCRIPT,
-  TAG_OPTION,
-  TAG_PARAM,
-  TAG_PLAINTEXT,
-  TAG_Q,
-  TAG_RP,
-  TAG_RT,
-  TAG_RUBY,
-  TAG_SAMP,
-  TAG_SCRIPT,
-  TAG_SMALL,
-  TAG_SOURCE,
-  TAG_SPAN,
-  TAG_STRONG,
-  TAG_STYLE,
-  TAG_SUB,
-  TAG_SUP,
-  TAG_TABLE,
-  TAG_TBODY,
-  TAG_TD,
-  TAG_TEMPLATE,
-  TAG_TEXTAREA,
-  TAG_TFOOT,
-  TAG_TH,
-  TAG_THEAD,
-  TAG_TIME,
-  TAG_TITLE,
-  TAG_TR,
-  TAG_TRACK,
-  TAG_U,
-  TAG_VAR,
-  TAG_WBR,
-  TAG_XMP,
-} from './const'
-
-const inline: TagHandler = {
-  collapsesInnerWhiteSpace: true,
-  spacing: NO_SPACING,
-  isInline: true,
-}
-
-const selfClosing: TagHandler = {
-  isSelfClosing: true,
-  spacing: NO_SPACING,
-}
-
-const selfClosingInline: TagHandler = {
-  ...selfClosing,
-  collapsesInnerWhiteSpace: true,
-  isInline: true,
-}
-
-const nonNesting: TagHandler = {
-  isNonNesting: true,
-  spacing: NO_SPACING,
-}
-
-const inertNonNesting: TagHandler = {
-  ...nonNesting,
-  excludesTextNodes: true,
-}
-
-/**
- * Parser-only tag metadata. This intentionally omits Markdown enter/exit
- * callbacks so text-only entrypoints do not import the Markdown renderer table.
- */
-export const tagMetadata: Record<number, TagHandler> = {
-  [TAG_HEAD]: {
-    spacing: NO_SPACING,
-    collapsesInnerWhiteSpace: true,
-  },
-  [TAG_SCRIPT]: {
-    isNonNesting: true,
-    excludesTextNodes: true,
-  },
-  [TAG_STYLE]: inertNonNesting,
-  [TAG_TITLE]: {
-    ...nonNesting,
-    collapsesInnerWhiteSpace: true,
-  },
-  [TAG_META]: {
-    ...selfClosing,
-    collapsesInnerWhiteSpace: true,
-  },
-  [TAG_BR]: selfClosingInline,
-  [TAG_STRONG]: inline,
-  [TAG_B]: inline,
-  [TAG_EM]: inline,
-  [TAG_I]: inline,
-  [TAG_DEL]: inline,
-  [TAG_SUB]: inline,
-  [TAG_SUP]: inline,
-  [TAG_INS]: inline,
-  [TAG_CODE]: inline,
-  [TAG_A]: inline,
-  [TAG_IMG]: selfClosingInline,
-  [TAG_THEAD]: {
-    spacing: TABLE_ROW_SPACING,
-    excludesTextNodes: true,
-  },
-  [TAG_TR]: {
-    spacing: TABLE_ROW_SPACING,
-    excludesTextNodes: true,
-  },
-  [TAG_TH]: {
-    collapsesInnerWhiteSpace: true,
-    spacing: NO_SPACING,
-  },
-  [TAG_TD]: {
-    collapsesInnerWhiteSpace: true,
-    spacing: NO_SPACING,
-  },
-  [TAG_SPAN]: inline,
-  [TAG_LABEL]: inline,
-  [TAG_BUTTON]: {
-    collapsesInnerWhiteSpace: true,
-    isInline: true,
-  },
-  [TAG_TBODY]: {
-    spacing: NO_SPACING,
-    excludesTextNodes: true,
-  },
-  [TAG_TFOOT]: {
-    spacing: TABLE_ROW_SPACING,
-    excludesTextNodes: true,
-  },
-  [TAG_KBD]: inline,
-  [TAG_LINK]: selfClosingInline,
-  [TAG_AREA]: selfClosingInline,
-  [TAG_BASE]: selfClosingInline,
-  [TAG_COL]: selfClosing,
-  [TAG_EMBED]: selfClosing,
-  [TAG_INPUT]: selfClosingInline,
-  [TAG_KEYGEN]: selfClosingInline,
-  [TAG_PARAM]: selfClosing,
-  [TAG_SOURCE]: selfClosing,
-  [TAG_TRACK]: selfClosing,
-  [TAG_WBR]: selfClosingInline,
-  [TAG_TEXTAREA]: nonNesting,
-  [TAG_OPTION]: nonNesting,
-  [TAG_IFRAME]: nonNesting,
-  [TAG_TEMPLATE]: inertNonNesting,
-  [TAG_ABBR]: inline,
-  [TAG_MARK]: inline,
-  [TAG_Q]: inline,
-  [TAG_SAMP]: inline,
-  [TAG_SMALL]: inline,
-  [TAG_NOSCRIPT]: {
-    excludesTextNodes: true,
-    spacing: NO_SPACING,
-  },
-  [TAG_DATALIST]: inertNonNesting,
-  [TAG_NOFRAMES]: nonNesting,
-  [TAG_XMP]: nonNesting,
-  [TAG_PLAINTEXT]: nonNesting,
-  [TAG_U]: inline,
-  [TAG_CITE]: inline,
-  [TAG_DFN]: inline,
-  [TAG_VAR]: inline,
-  [TAG_TIME]: inline,
-  [TAG_BDO]: inline,
-  [TAG_RUBY]: inline,
-  [TAG_RT]: inline,
-  [TAG_RP]: inline,
-  [TAG_FIGCAPTION]: inline,
-  [TAG_TABLE]: {},
-}
